+++ sdk/lib/_internal/js_shared/lib/js_interop_patch.dart
@@ -486,6 +486,121 @@ extension StringToJSString on String {
   JSString get toJS => JSString._(JSStringType(this));
 }
 
+// -----------------------------------------------------------------------------
+// JSArray<JSNumber> <-> List<num>
+@patch
+extension JSArrayOfNumberToList on JSArray<JSNumber> {
+  @patch
+  @pragma('dart2js:prefer-inline')
+  List<double> get toDartOfDouble =>
+      this as List<double> ?? (this as List).cast<double>();
+
+  @patch
+  @pragma('dart2js:prefer-inline')
+  List<int> get toDartOfInt => this as List<int> ?? (this as List).cast<int>();
+}
+
+@patch
+extension ListOfNumberToJSArray on List<num> {
+  @patch
+  @pragma('dart2js:prefer-inline')
+  JSArray<JSNumber> get toJS => (this as JSAny).isA<JSArray>()
+      ? this as JSArray<JSNumber>
+      : [for (var value in this) value.toJS].toJS;
+}
+
+// -----------------------------------------------------------------------------
+// JSArray<JSString> <-> List<String>
+@patch
+extension JSArrayOfStringToList on JSArray<JSString> {
+  @patch
+  @pragma('dart2js:prefer-inline')
+  List<String> get toDart =>
+      this as List<String> ?? (this as List).cast<String>();
+}
+
+@patch
+extension ListOfStringToJSArray on List<String> {
+  @patch
+  JSArray<JSString> get toJS => (this as JSAny).isA<JSArray>()
+      ? this as JSArray<JSString>
+      : [for (var value in this) value.toJS].toJS;
+}
+
+// -----------------------------------------------------------------------------
+// JSArray<JSBoolean> <-> List<bool>
+@patch
+extension JSArrayOfBooleanToList on JSArray<JSBoolean> {
+  @patch
+  List<bool> get toDart => this as List<bool> ?? (this as List).cast<bool>();
+}
+
+@patch
+extension ListOfBoolToJSArray on List<bool> {
+  @patch
+  JSArray<JSBoolean> get toJS => (this as JSAny).isA<JSArray>()
+      ? this as JSArray<JSBoolean>
+      : [for (var value in this) value.toJS].toJS;
+}
+
+// -----------------------------------------------------------------------------
+// JSArray<JSNumber?> <-> List<num?>
+@patch
+extension JSArrayOfNullableNumberToList on JSArray<JSNumber?> {
+  @patch
+  @pragma('dart2js:prefer-inline')
+  List<double?> get toDartOfDouble =>
+      this as List<double?> ?? (this as List).cast<double?>();
+
+  @patch
+  @pragma('dart2js:prefer-inline')
+  List<int?> get toDartOfInt =>
+      this as List<int?> ?? (this as List).cast<int?>();
+}
+
+@patch
+extension ListOfNullableNumberToJSArray on List<num?> {
+  @patch
+  @pragma('dart2js:prefer-inline')
+  JSArray<JSNumber?> get toJS => (this as JSAny).isA<JSArray>()
+      ? this as JSArray<JSNumber?>
+      : [for (var value in this) value?.toJS].toJS;
+}
+
+// -----------------------------------------------------------------------------
+// JSArray<JSString?> <-> List<String?>
+@patch
+extension JSArrayOfNullableStringToList on JSArray<JSString?> {
+  @patch
+  @pragma('dart2js:prefer-inline')
+  List<String?> get toDart =>
+      this as List<String?> ?? (this as List).cast<String?>();
+}
+
+@patch
+extension ListOfNullableStringToJSArray on List<String?> {
+  @patch
+  JSArray<JSString?> get toJS => (this as JSAny).isA<JSArray>()
+      ? this as JSArray<JSString?>
+      : [for (var value in this) value?.toJS].toJS;
+}
+
+// -----------------------------------------------------------------------------
+// JSArray<JSBoolean?> <-> List<bool?>
+@patch
+extension JSArrayOfNullableBooleanToList on JSArray<JSBoolean?> {
+  @patch
+  List<bool?> get toDart => this as List<bool?> ?? (this as List).cast<bool?>();
+}
+
+@patch
+extension ListOfNullableBoolToJSArray on List<bool?> {
+  @patch
+  JSArray<JSBoolean?> get toJS => (this as JSAny).isA<JSArray>()
+      ? this as JSArray<JSBoolean?>
+      : [for (var value in this) value?.toJS].toJS;
+}
+
 @patch
 extension JSAnyOperatorExtension on JSAny? {
   @patch