+++ sdk/lib/js_interop/js_interop.dart
@@ -2046,6 +2046,202 @@ extension StringToJSString on String {
   external JSString get toJS;
 }
 
+/// Conversions from `JSArray<JSNumber>` directly to Dart lists.
+@Since('3.13')
+extension JSArrayOfNumberToList on JSArray<JSNumber> {
+  /// Converts this to a [List] of [double]s.
+  ///
+  /// **Note:** Depending on the platform and which option is most efficient,
+  /// this may copy the original list, it may return a view over it, or it may
+  /// return it as-is with a different static type. Which one happens under
+  /// which circumstance is not guaranteed. For compatibility across platforms,
+  /// avoid modifying the original or returned list without explicitly copying
+  /// it yourself.
+  external List<double> get toDartOfDouble;
+
+  /// Converts this to a [List] of [int]s.
+  ///
+  /// **Note:** Depending on the platform and which option is most efficient,
+  /// this may copy the original list, it may return a view over it, or it may
+  /// return it as-is with a different static type. Which one happens under
+  /// which circumstance is not guaranteed. For compatibility across platforms,
+  /// avoid modifying the original or returned list without explicitly copying
+  /// it yourself.
+  ///
+  /// Calling this on a list of non-integer values will produce a [TypeError],
+  /// but that error may be thrown either on calling this method or on accessing
+  /// the values of the list.
+  external List<int> get toDartOfInt;
+}
+
+/// Conversions from `List<num>` directly to `JSArray<JSNumber>`.
+@Since('3.13')
+extension ListOfNumberToJSArray on List<num> {
+  /// Converts this to a [JSArray] of [JSNumber]s.
+  ///
+  /// **Note:** Depending on the platform and which option is most efficient,
+  /// this may copy the original list, it may return a view over it, or it may
+  /// return it as-is with a different static type. Which one happens under
+  /// which circumstance is not guaranteed. For compatibility across platforms,
+  /// avoid modifying the original or returned list without explicitly copying
+  /// it yourself.
+  external JSArray<JSNumber> get toJS;
+}
+
+/// Conversions from `JSArray<JSString>` directly to `List<String>`.
+@Since('3.13')
+extension JSArrayOfStringToList on JSArray<JSString> {
+  /// Converts this to a [List] of [String]s.
+  ///
+  /// **Note:** Depending on the platform and which option is most efficient,
+  /// this may copy the original list, it may return a view over it, or it may
+  /// return it as-is with a different static type. Which one happens under
+  /// which circumstance is not guaranteed. For compatibility across platforms,
+  /// avoid modifying the original or returned list without explicitly copying
+  /// it yourself.
+  external List<String> get toDart;
+}
+
+/// Conversions from `List<String>` directly to `JSArray<JSString>`.
+@Since('3.13')
+extension ListOfStringToJSArray on List<String> {
+  /// Converts this to a [JSArray] of [JSString]s.
+  ///
+  /// **Note:** Depending on the platform and which option is most efficient,
+  /// this may copy the original list, it may return a view over it, or it may
+  /// return it as-is with a different static type. Which one happens under
+  /// which circumstance is not guaranteed. For compatibility across platforms,
+  /// avoid modifying the original or returned list without explicitly copying
+  /// it yourself.
+  external JSArray<JSString> get toJS;
+}
+
+/// Conversions from `JSArray<JSBoolean>` directly to `List<bool>`.
+@Since('3.13')
+extension JSArrayOfBooleanToList on JSArray<JSBoolean> {
+  /// Converts this to [List] of [bool]s.
+  ///
+  /// **Note:** Depending on the platform and which option is most efficient,
+  /// this may copy the original list, it may return a view over it, or it may
+  /// return it as-is with a different static type. Which one happens under
+  /// which circumstance is not guaranteed. For compatibility across platforms,
+  /// avoid modifying the original or returned list without explicitly copying
+  /// it yourself.
+  external List<bool> get toDart;
+}
+
+/// Conversions from `List<bool>` directly to `JSArray<JSBoolean>`.
+@Since('3.13')
+extension ListOfBoolToJSArray on List<bool> {
+  /// Converts this to a [JSArray] of [JSBoolean]s.
+  ///
+  /// **Note:** Depending on the platform and which option is most efficient,
+  /// this may copy the original list, it may return a view over it, or it may
+  /// return it as-is with a different static type. Which one happens under
+  /// which circumstance is not guaranteed. For compatibility across platforms,
+  /// avoid modifying the original or returned list without explicitly copying
+  /// it yourself.
+  external JSArray<JSBoolean> get toJS;
+}
+
+/// Conversions from `JSArray<JSNumber?>` directly to Dart lists.
+@Since('3.13')
+extension JSArrayOfNullableNumberToList on JSArray<JSNumber?> {
+  /// Converts this to a [List] of nullable [double]s.
+  ///
+  /// **Note:** Depending on the platform and which option is most efficient,
+  /// this may copy the original list, it may return a view over it, or it may
+  /// return it as-is with a different static type. Which one happens under
+  /// which circumstance is not guaranteed. For compatibility across platforms,
+  /// avoid modifying the original or returned list without explicitly copying
+  /// it yourself.
+  external List<double?> get toDartOfDouble;
+
+  /// Converts this to a [List] of nullable [int]s.
+  ///
+  /// **Note:** Depending on the platform and which option is most efficient,
+  /// this may copy the original list, it may return a view over it, or it may
+  /// return it as-is with a different static type. Which one happens under
+  /// which circumstance is not guaranteed. For compatibility across platforms,
+  /// avoid modifying the original or returned list without explicitly copying
+  /// it yourself.
+  ///
+  /// Calling this on a list of non-integer values will produce a [TypeError],
+  /// but that error may be thrown either on calling this method or on accessing
+  /// the values of the list.
+  external List<int?> get toDartOfInt;
+}
+
+/// Conversions from `List<num>` directly to `JSArray<JSNumber>`.
+@Since('3.13')
+extension ListOfNullableNumberToJSArray on List<num?> {
+  /// Converts this to a [JSArray] of nullable [JSNumber]s.
+  ///
+  /// **Note:** Depending on the platform and which option is most efficient,
+  /// this may copy the original list, it may return a view over it, or it may
+  /// return it as-is with a different static type. Which one happens under
+  /// which circumstance is not guaranteed. For compatibility across platforms,
+  /// avoid modifying the original or returned list without explicitly copying
+  /// it yourself.
+  external JSArray<JSNumber?> get toJS;
+}
+
+/// Conversions from `JSArray<JSString?>` directly to `List<String?>`.
+@Since('3.13')
+extension JSArrayOfNullableStringToList on JSArray<JSString?> {
+  /// Converts this to a [List] of nullable [String]s.
+  ///
+  /// **Note:** Depending on the platform and which option is most efficient,
+  /// this may copy the original list, it may return a view over it, or it may
+  /// return it as-is with a different static type. Which one happens under
+  /// which circumstance is not guaranteed. For compatibility across platforms,
+  /// avoid modifying the original or returned list without explicitly copying
+  /// it yourself.
+  external List<String?> get toDart;
+}
+
+/// Conversions from `List<String>` directly to `JSArray<JSString>`.
+@Since('3.13')
+extension ListOfNullableStringToJSArray on List<String?> {
+  /// Converts this to a [JSArray] of nullable [JSString]s.
+  ///
+  /// **Note:** Depending on the platform and which option is most efficient,
+  /// this may copy the original list, it may return a view over it, or it may
+  /// return it as-is with a different static type. Which one happens under
+  /// which circumstance is not guaranteed. For compatibility across platforms,
+  /// avoid modifying the original or returned list without explicitly copying
+  /// it yourself.
+  external JSArray<JSString?> get toJS;
+}
+
+/// Conversions from `JSArray<JSBoolean?>` directly to `List<bool?>`.
+@Since('3.13')
+extension JSArrayOfNullableBooleanToList on JSArray<JSBoolean?> {
+  /// Converts this to [List] of nullable [bool]s.
+  ///
+  /// **Note:** Depending on the platform and which option is most efficient,
+  /// this may copy the original list, it may return a view over it, or it may
+  /// return it as-is with a different static type. Which one happens under
+  /// which circumstance is not guaranteed. For compatibility across platforms,
+  /// avoid modifying the original or returned list without explicitly copying
+  /// it yourself.
+  external List<bool?> get toDart;
+}
+
+/// Conversions from `List<bool>` directly to `JSArray<JSBoolean>`.
+@Since('3.13')
+extension ListOfNullableBoolToJSArray on List<bool?> {
+  /// Converts this to a [JSArray] of nullable [JSBoolean]s.
+  ///
+  /// **Note:** Depending on the platform and which option is most efficient,
+  /// this may copy the original list, it may return a view over it, or it may
+  /// return it as-is with a different static type. Which one happens under
+  /// which circumstance is not guaranteed. For compatibility across platforms,
+  /// avoid modifying the original or returned list without explicitly copying
+  /// it yourself.
+  external JSArray<JSBoolean?> get toJS;
+}
+
 /// General-purpose JavaScript operators.
 ///
 /// Indexing operators (`[]`, `[]=`) should be declared through operator