+++ CHANGELOG.md
@@ -112,6 +112,21 @@ To learn more about the feature, check out the
   the runtime types of `JSFunction` and `JSExportedDartFunction` do not change.
   For more details, see SDK issue [#54557][].
 
+- **Breaking change**:
+  Added `JSArrayOfStringToList`, `JSArrayOfBooleanToList`,
+  `JSArrayOfNullableStringToList`, and `JSArrayOfNullableBooleanToList`. These
+  eagerly and efficiently convert the contents of the array to Dart values, but
+  they will shadow previous calls to `JSArray<T extends JSAny?>.toDart` which
+  may cause build errors in code that expected the resulting array to contain `T
+  extends JSAny?` rather than Dart values.
+
+- Added `JSArrayOfNumberToList`, `ListOfNumberToJSArray`,
+  `ListOfStringToJSArray`, `ListOfBoolToJSArray`,
+  `JSArrayOfNullableNumberToList`, `ListOfNullableNumberToJSArray`,
+  `ListOfNullableStringToJSArray`, and `ListOfNullableBoolToJSArray`. These make
+  it easier and more efficient to convert between JS and Dart arrays of
+  primitives.
+
 [#54557]: https://github.com/dart-lang/sdk/issues/54557
 
 ### Tools